+++ scripts/module.js
@@ -21,6 +21,7 @@ import { scene } from './utility/scene.js';
 //CONFIG.debug.hooks = true;
 
 export let socket;
+
 Hooks.once('init', async function() {
 	
 });
@@ -39,6 +40,36 @@ Hooks.once('ready', async function() {
 	
 });
 
+Hooks.on("getCompendiumDirectoryEntryContext", (html, options) => {
+	options.push({
+	  name: game.i18n.localize("sdnd.compendium.entry.context.exportThumbnails"),
+	  icon: '<i class="fa-solid fa-camera-retro"></i>',
+	  callback: async function(li) {
+		debugger;
+		const fullyQualifiedPack = $(li).data("pack");
+		if (!fullyQualifiedPack){
+			return;
+		}
+		const moduleId = fullyQualifiedPack.split('.').shift();
+		await scene.packModuleThumbnails(moduleId);
+	  },
+	  condition: li => {
+		debugger;
+		const fullyQualifiedPack = $(li).data("pack");
+		if (!fullyQualifiedPack){
+			return false;
+		}
+		const moduleId = fullyQualifiedPack.split('.').shift();
+		if (!moduleId){
+			return;
+		}
+		if (moduleId.startsWith('sdnd-') || moduleId.startsWith('stroud-')){
+			return true;
+		}
+		return false;
+	  },
+	});
+  });
 Hooks.on("getSceneDirectoryFolderContext", (html, options) => {
 	options.push({
 	  name: game.i18n.localize("sdnd.scenes.folder.context.regenerateThumbs"),
@@ -48,19 +79,29 @@ Hooks.on("getSceneDirectoryFolderContext", (html, options) => {
 		if (!folderID){
 			return;
 		}
-		scene.regenerateThumbnails(folderID);
+		let folder = await game.folders.get(folderID);
+		let thumbCount = await scene.regenerateThumbnails(folder);
+		if (thumbCount > 0) {
+			ui.notifications.notify(`Regenerated ${thumbCount} thumbnail image${(thumbCount > 0 ? 's' : '')}.`);
+		}
 	  },
 	  condition: li => {
 		let folderID = $(li).closest("li").data("folderId");
 		if (!folderID) {
 			return false;
 		}
-		// get scene count
-		let sceneCount = game.scenes.filter(s => s.folder?.id == folderID)?.length ?? 0;
-		if (sceneCount > 0){
-			return true;
+		let folder = game.folders.get(folderID);
+		if (!folder) {
+			return false;
 		}
-		return false;
+		let sceneCount = game.scenes.filter(s => s.folder?.id == folder.id)?.length ?? 0;
+		if (sceneCount == 0 && folder.children?.length == 0) {
+			return false;
+		}
+		if (sceneCount?.length > 10 || folder.children?.length > 10){
+			return false;
+		}
+		return true;
 	  },
 	});
   });